+++ frontend/vendor-web/src/pages/CollaborationRequest/components/EventForm.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { setEventForm } from '../store/formsSlice';
-import { setStep } from '../store/stepSlice';
+import { setEventForm } from '../../../store/formsSlice';
+import { setStep } from '../../../store/stepSlice';
 import {
     Box,
     FormControl,
