--- frontend/vendor-web/src/pages/EventsRoot.js
@@ -1,13 +1,0 @@
-import { Outlet } from "react-router-dom";
-import EventsNavigation from '../components/EventsNavigation';
-
-function EventsRootLayout () {
-    return (
-        <>
-            <EventsNavigation />
-            <Outlet />
-        </>
-    );
-}
-
-export default EventsRootLayout;
