+++ frontend/vendor-web/src/pages/Home/components/Carousel.jsx
