+++ frontend/vendor-web/src/pages/Home/components/Game.jsx
