+++ frontend/vendor-web/src/pages/Home/components/slickBtn.jsx
