--- frontend/vendor-web/src/pages/NewEvent.js
@@ -1,5 +1,0 @@
-function NewEventPage() {
-    return <>This is NewEventPage</>;
-}
-
-export default NewEventPage;
